plat_stm32f411: i2c: support interrupted transactions
The slave must still do one RX followed by one or more TX, but these can
be divided up across multiple transactions.  To simplify the
implementation, the initial RX must still happen in a single I2C packet.
(The host interface protocol allows the master to interrupt reads but
not writes.)

This requires a few tweaks to the driver's state machine: mostly
shifting things around so ACKs are still enabled if the transaction ends
right after the RX, and letting the client queue up TX buffers when in
that state.

Change-Id: I68bb5759593641e78a5f3cbbc268ddc079fc1167
Signed-off-by: Greg Hackmann <ghackmann@google.com>