gpio: Seperate gpioConfig into 3 separate functions (in/out/alt)
It was impossible to configure an active low pin for output (with it
initially unasserted) without causing a glitch. Now you specify the
initial value for the output when you configure the gpio pin. The same
is also true for pins specified as Alternate Function. You specify the
function when configuring the pin.

Change-Id: I629dcf79be3b8e7725717455be543c7741e15128