spi: Fix access of free'd memory with multiple spiMasterRxTx calls
spiMasterRxTx allocates memory off the heap and passes it to spi via
spiRequest, which sets mStmSpiDevs[busId].base to that memory. After
the transfer is complete, the memory is freed but
mStmSpiDevs[busId].base still points to it. On a second call to
spiMasterRxTx it allocates new memory and passes it to spiRequest, but
spiRequest only assigns mStmSpiDevs[busId].base if it was originally
NULL, so it doesn't overwrite the pointer and it instead points to the
now freed (or allocated by something else at this point) memory.

Add a release function that assigns mStmSpiDevs[busId].base back to NULL

Change-Id: If765f6e1abc16a438240fe448348222984da8132