spi: add slave API
The slave API is similar to the master API but slightly more stateful,
since the slave can be enabled even while there are no buffers queued up
for transmission.

To simplify the SPI driver, the common SPI code is responsible for
worrying about the SPI state machine.  Instead the SPI driver must
implement spiSlaveIdle() that puts the device in an idle state (i.e.,
discard incoming bytes and respond with the preamble) when requested.

Change-Id: I42319d07925cacf1798be317be308de57df814eb
Signed-off-by: Greg Hackmann <ghackmann@google.com>