inc: add chained ISR helper
A simple, generic helper for chaining together ISRs for a shared IRQ.
ISRs can be efficiently added or removed from a chain, even from inside
an interrupt context.  The shared IRQ is automatically enabled or
disabled as ISRs are added and removed from the chain.

The platform is responsible for declaring the appropriate
ChainedInterrupt structures, providing callbacks to enable/disable the
underlying IRQ, and calling dispatchIsr() in the corresponding IRQ
handler.

Change-Id: I5fa36dd4c4092d4932c57cd8f4ebd8b78f558739
Signed-off-by: Greg Hackmann <ghackmann@google.com>