spi: deal with rare case where first RXNE happens before (or with) TXE
This causes the last byte to not be transmitted. To deal with this case
check to make sure we have received state->words bytes if rxBuf is set
and sent state->words byte if txBuf is set (require both if we are sending
and receiving at the same time). Check in both stmSpiTxe and stmSpiRxne
so we will call stmSpiDone() at the earliest opportunity.

Change-Id: I2fe604e08d748e767403b9cbe94de63ca506f975