stm32f4xx: spi: make some configuration data variant-specific
Several parts of the SPI configuration are tied to how the GPIOs are
configured, which is specific to the board layout.  Split this data off
into a separate struct StmSpiPinCfg and require variant support code to
supply this information.

Also supply the appropriate support code for the pike board variant.

Change-Id: Ia5ff9323692b649a8d3e9b1af2fa231c686fd600
Signed-off-by: Greg Hackmann <ghackmann@google.com>