stm32f4xx: spi: use DMA
DMA for SPI masters is straightforward, since the master drives the
clock: switching buffers just means reprogramming and re-enabling the
DMA controller.

DMA for SPI slaves is trickier because the master will keep trying to
read bytes while the OS is occupied with setting up the next transfer.
Restarting DMA at the wrong point can also cause the first word in the
packet to be corrupted.  Working around this with double-buffering is
impractical since the DMA controller does not allow setting the sizes of
the buffers independently or changing the size on-the-fly.

Experimentation found the most reliable way to handle this is to use DMA
for buffers, and interrupts to transmit preambles (including during gaps
between DMA transfers, when the master might keep requesting bytes).  In
practice the CPU doesn't need to keep handling the TXE interrupt to do
this; servicing it once then disabling TXEIE is enough to make the SPI
controller repeat the same word.

Restarting DMA is then just a matter of resetting RXNE and TXE, so the
FIFO is in a sane state just before handing everything back to DMA.

Change-Id: I3c5ef12690d7e1464925c206507118fbd2ee93a2
Signed-off-by: Greg Hackmann <ghackmann@google.com>