stm32f4xx: spi: always enable RX channel on master
Leaving the RX channel disabled can put the SPI and/or DMA controllers
in a state where the SPI FIFO is supposedly empty (RXNE = 0) yet the DMA
controller "receives" one of the discarded bytes at the start of the
next transmission.

Work around this by unconditionally enabling the RX channel when in
master mode.  In cases where we don't actually care about the data
(rxBuf = NULL) then RX into a throwaway one-word buffer with MINC = 0.

Oddly, this workaround only appears to be necessary in master mode, and
using it in slave mode actually creates more problems than it solves
(sometimes the first TX byte is dropped).

Bug: 22906229

Change-Id: I4868f6c2a6c031f109bc8f27ad520debe211fd68
Signed-off-by: Greg Hackmann <ghackmann@google.com>