spi: do not copy input packet list
Profiling showed the heapAlloc() calls inside spi{Master,Slave}RxTx()
are a performance hotspot.  Eliminate these allocations by requiring the
caller to temporarily transfer ownership of the packet lists over to
SPI.

This is an API-breaking change, and the caller must be sure not to pass
lists allocated on the stack.

Also remove the helper functions that depended on stack-allocated lists.

Bug: 23256798

Change-Id: I53bd5a4926d53090dd248d1a67cefc33a4947fea
Signed-off-by: Greg Hackmann <ghackmann@google.com>