security: add RSA keys and infrastructure to get them
Keys live in bootloader, and it gives API to access them
Each device has 3 or 4 roots of trust:
 - Google
 - Vendor
 - Device
 - [optional] Google Debug key

All keys are actually owned by google, but vendor and device keys will be used to sign a
vendor's set of keys, whcih then become vendor's keys to sign apps with.
Debug key is only included in debug builds. Its private key is available in misc/ for
testing

Change-Id: I329248340ced87f990d2e086f1592ce983e98283