seos: refactor how events are freed & misc fixes
In preparation for apps being able to enqueue events, we need a way for them to free them.
This change sens a special event to apps that this is needed, and the data is the event
the app needs to free. This causes a large restructuring, as to avoid wasting space we now
use tagged pointers to multiplex the "free callback" and "free app's TID" in the event.
This also was a reminder to limit the range of TIDs, which was done. Also some includes
That depended on tothe rincludes did not include them. Fixed

Change-Id: Ib2e121d5639e6bbc0e5502a8ec2e8904ff2c1437