hostintf: Send preamble byte as first byte and support retransmits
Work around issue with spi dma startup by always sending a preamble
byte as the first byte of each packet (when starting up dma with
the AP already transmitting the first byte sent can sometimes either
be dropped, or merged with the previous byte being sent). By sending
a preamble byte as the first byte neither case will cause an issue.

When receiving a message with the same sequence number as the previous
message, assume the AP lost the previous message and retransmit what
we already transmitted for that sequence number.

Change-Id: Id1362bc3ec4625800b2547ac4e161934474d879c
Signed-off-by: Ben Fennema <fennema@google.com>