make: implement checks for bad functions
If a function is found in the binary that should never ever be used in
shipping code For performance, security, or another reason, bail out with
an error. The list of bad functions is assembled in the Makefiles in the
BADWORDS variable. Each entry is either just a function name or in the
format aaa=bbb where aaa is the function name and bbb is a descriptive
message to print in case it is seen. For C-M4F we warn on slow softfloat
functions. Eventualy we'll warn on long divide too, but only once we've
removed the ones we have alrady.

Change-Id: I9c21e1dc2c546a2dd67425f807dadd12a42fa187