bmi160: prevent multiple time syncs
The active_data_sensor_cnt variable was uninitialised and
is needed to make sure that only a single time sync event
is scheduled every 100ms. Previously when a fusion algorithm
like gravity was started, two time sync events would be
scheduled every 100ms, one for each of the two enabled
sensors. This was causing unintended wakeups.

Change-Id: I81c756ae1a6866d58f6ac769fcc21e567c350f7e