bmp280: only write control register when it changes
If the BMP280 is performing a measurement, any I2C transaction to change
the power mode is stalled until the measurement is complete. Any
subsequent writes to the control register are ignored. This CL fixes a
race condition where an unnecessary write to the control register was
blocking a necessary write, causing the device to stay powered on.

Bug: 26837113
Change-Id: I5e9cd6ef8ab567498295273ebbb40b7f013e2d20
(cherry picked from commit e2386d755e68eb9ce7d6620daa5af1dd61033b01)