seos: add ability to retain events
Add is a mechanist to signal to the OS that the current "being handled" event
is not to be freed. Calling party takes on all responsibility for freeing it
later in whatever way it wishes. Only one app may do this. This is implemented
using a global in seos.c to avoid changing app ABI so late in the game.