sensors: allow user events to provide marshalling callbacks
Any uploaded nanoapp may act as a sensor and send sensor data. Sensor data may
be in a format that NanohubOS will not understand (nanoapp might be from 2017,
NanohubOS could have been built in 2015). We solve this by allowing any custom
user events to provide marshalling callbacks. HostIF, if it so chooses, will
use those to marshall events for the wire. The mashalling func is also
responsible for freeing the passed-in event using osFreeRetainedEvent()
ALWAYS

Change-Id: I7467e8304bead7f808abcbd94cd1492f9373ed2f