nanotool: Fix bug in error reporting.
As written, the "else if" clause would never be entered because it
was identical to the "if" clause.  The original "if" clause was
incorrect, as it's only when write() returns -1 that it is
documented to fill in errno.  We fix this.

Note this bug only affected the error case, and still had us
succeeding iff the write() fully succeeded.  So this bug was
of very limited impact.

Change-Id: I78f516141e86d0a15b2b4aae678d9fbe4a7b84d8