nanohub: Block HAL read threads when lock file present
Support factory calibration, low-level debug, etc. using direct access
to the device file while the Android runtime is up without interference
from the sensors HAL or context hub HAL by using a lock file. When the
HALs notice that this lock file is present, their read threads will
block until the file is removed, rather than reading from their
respective device files. This avoids the race between the read threads
and a command-line utility that are both attempting to read events from
a device file with no inherent multi-client support.

This lock file is created and deleted by nanotool (or potentially some
other utility in the future), assuming it is run with sufficient
privileges to access the directory.

The current implementation sacrifices true inter-process synchronization
for lower overhead in the read threads, meaning that nanotool needs to
use a time delay to have reasonable assurance that the HAL threads have
noticed the creation of the lock file.

Bug: 27569995
Change-Id: I692acc4ca773151fbd677ac1d71f526c598b4e5f