nanoapp_cmd: Fix compiler warning.
We note that 'sig' is unused as a parameter.  Out of paranoia, we
also assert that it's SIGINT, so if we use this handler for multiple
signals in the future and forget to update it, we'll find out.

Change-Id: I44e7ac3dc85a748db46b33be1dda51b0e5de140a
(cherry picked from commit 2041bc7bdb8f124dcc207da884dd414250b11b2e)