contexthubhal: Fix default case.
C allows labels within seemingly odd places, including within a
'switch' statement.  Thus, the misspelled 'dfault' was just a
label without a corresponding 'goto', and we never logged an
unknown message type.

We fix this typo (and work towards turning on compiler warnings
as errors, to avoid these issues in the future).

Change-Id: I585b3dea7a97ee22b3a74635a4bd81bfad4ba457