contexthubhal: Fix compiler warning, minor rework.
While fixing a compiler warning about a comparison between signed
and unsigned types, we rework our constants a little.  We no longer
hardcode 1 for the length required to store the packet length.
And instead of adding together our event id, app name, and packet
length lengths in multiple areas, we create a constant for it.

We also give more detailed log messages in error cases, and avoid
reading uninitialized memory from the 'header' if we failed to
read as many bytes as we expected.

Change-Id: I3ce246f9c4bf08f590bb2d250005d5fd7f93bbc5