Add encode support for deltaTime in hostIntf
This is the hostIntf side change to support the encoding/decoding of deltaTime.

Previously the deltaTime is always in nano second. Given 32bits, it is limited
to ~4sec. Now by making the LSB of deltaTime to indicate whethear the deltaTime
is in nanosecond or nanosceond/1024, it will allow upto ~4000sec. This is useful
for certian nanoapps.

Bug: 27489503

Change-Id: Ib299a9c11c2d48c6a4d04a621e72a9b02f6b0ec2