Add decode support for deltaTime in hubconnection
This is the hubconnection side change to support the encoding/decodiing of
deltaTime.

Previously the deltaTime is always in nano second. Given 32bits, it is limited
to ~4sec. Now by making the LSB of deltaTime to indicate whethear the deltaTime
is in nanosecond or nanosceond/1024, it will allow upto ~4000sec. This is useful
for certian nanoapps.

Bug: 27489503
Change-Id: I06b42aca18cb20e5f4403cd9385331854cedff33