nanoapp_postprocess: Make "make clean" idempotent
We use the '-f' with the "rm" command so there is no failure if
the app does not exist.  This allows the user or a script to call
"make clean" an arbitrary number of times without failure, or even
just a single time with the tree in an arbitrary state.

Change-Id: Icfa765a1a68fc3f1ec76a560c06b22907269f543