nanoapp_postprocess: Use Android build system
We embrace more of the Android build system, instead of
entirely using a parallel system of our own creation.
This change has the advantage that the user (and
eventually, a build server) who is going through
the Android makefile no longer needs to alter their path
to be able to access the nanoapp_postprocess executable.

We leave the Makefile, which remains a parallel build
path and thus continues to be subject to diverging from
the Android build system.  There are concerns over how
some customers may want to build nanoapps outside of
the Android build system, so potential unification of
these paths is much more complex and well beyond the
scope of this CL.

Bug: 27997616
Change-Id: If2a00d9ef5e99a2cac849fba484490cf69f5df72