contexthub: Remove leading tabs from C code
We remove tabs which start lines of C code that we own.  There
are many other places in our source where we're using hard-coded
tabs, but these seem the most critical, where they can confuse
readers of the code who have different tab settings.

In arm_math.h, we also change a couple indentations to be
consistent with the style in that header.  While it's a
different style than ours, we value minimizing our changes.

Change-Id: I1fe8e166c10c55b3bce178c799b271fca6e2642f