nanohub: firmware: Introduce safe event send
Event send method (osEnqueueEvt) is not safe;
in order to prevent memory leak, its return value
must always be checked, and memory deallocated, if needed.

It is not practical to change API at this time, since this
method is used in many contexts. Instead, let's define new method,
and this time do the safe handling w.r.t. memory leaks.

Bug: 28265099
Change-Id: Ib5ea03d5247ebc5a6cee743ae2b0460406e41b2c
Signed-off-by: Alexey Polyudov <apolyudov@google.com>