nanoapp: correctly set app version section
The section of the app version is currently set to .appversion. The
linker script is expecting .app_version. This results in the section
being optimized away due to gc-sections and all apps end up with a
version of zero (because this just happens to be the memory following
the location where the version number should be).

Bug: 28275786
Change-Id: I1f9c109d80aa2ae4963438b38662408f4d2d75cb