platform: Add support for explicit wake from sleep
On some platforms, handling an interrupt does not implicitly cause the
normal context to wake from the suspended platSleep() - this needs to be
done explicitly from the interrupt context. This change adds a new
platform API, platWake(), to support this model, and invokes it when
posting an event to the queue.

Bug: 28342647
Change-Id: I4b383968e73bf94b5cf9d464d964731bc2de3eec