hostIntf: re-enable interrupts once we resume reading
Mark all non-wakeup samples that are queued infront of wakeup samples as
wakeup so the kernel driver will take the wakelock and prevent the system
from suspending before there is time to get to the wakeup sample.

If the kernel driver runs out of buffers, it temporarily disables interrupts.
(this can occur during suspend and resume where userspace is suspended but
the kernel driver is running and will continue draining buffers from the nanohub
fifo)
Make sure to re-enable interrupts once the kernel driver resumes reading;
otherwise, the kernel driver can stop prior to fully draining the nanohub fifo.

Bug: 28470911

Change-Id: I98d0856ec2d890fcbdb4377b83f6067246403e49
Signed-off-by: Ben Fennema <fennema@google.com>