[WIP] In-Situ Gyroscope Calibration Algorithm.
Added initial gyro calibration files. Compiled, flashed, and produced
calibrations.

[05-12-2016] Added print out of floats. Verified correct calibration
funcitonality for the gyro cal code.

[05-17-2016] Removed typedef's and used 'struct' instead of 'typedef struct'.
Added printout of most recent calibration data, and calibration data that is
limited in frequency to every 2 hours. This is to help ensure that we gather
long term history of gyro calibrations (Dogfooding). Needed to use a simple
state machine to do the Dogfood reporting in bosch_bmi160.c. This was done to
ensure that the ASCII printing is properly throttled so that it does not
interfere with other system processing.

[05-18-2016] Put in fixes from code review, and adjusted style to blend with
Android code base.

[05-24-2016] Put in fixes from code review; minor style adjustments.

[05-31-2016] Cleared error in stillness logic. With retuning of the stillness
parameters, there was no impact on algorithm testing results.

[06-02-2016] Minor cleanup.

Bug: 28981599
Bug: 28635583

Change-Id: I1b852d6223d0539a1ecd93a6233be0e1622ade28