CodingStandard: Move under firmware.
This coding standard does not apply to the CHRE API.  Thus,
it's confusing to have it located at this higher level which
implies that it's relevant to such code.

We move this under the firmware directory so its clearer that
this is focused on the nanohub implementation.  This standard
still has issues and doesn't fit the current nanohub implementation
that well, but we'll leave modifications or removal of this
file to followup CLs.

Change-Id: Iec9aa0925e0d97d480dab564d60748e5596a7dc8