nanohub: bl: align default flash table to default chip
This fixes a crash condition that occurs on the variant nucleo (CHIP
STM32F411) when the shared memory is corrupted or full and the FW tries
to erase the flash. Due to wrong flash table, the firmware erases the
firmware itself and then crashes.
This is caused by an incorrect erase table that misses a flash block in
the code segment (normally 16K block + 64K block + 128K block defined
in the link script).
It erroneously erases the sector 5 (code) rather the sector 6 (shared).

The patch alignes the default flash table to the linker script of the
default CHIP STM32F411 specified in firmware/variant/variant.mk
Variants that use CHIPs other than the default one should provide a
specific flash table in variant.h.

Change-Id: I4d339e75c131632255838632d7fb1356dfaba684
Signed-off-by: William Hennebois <william.hennebois@st.com>
Signed-off-by: Antonio Borneo <antonio.borneo@st.com>