Refactor to eliminate taking the address of a packed member.
Bug: http://b/31532493

NanoHub::doRun() was previously taking the address of msg.hdr.app_name,
which isn't allowed, as it could possibly be misaligned. In fact,
looking at the data structure layout of the packed, nested structs,
app_name (a struct containing only a single uint64_t) will always be
packed immediately after a uint32_t. No matter what alignment we want to
apply to the main struct, we will never end up with an aligned uint64_t.
To address this, we merely switch to copying the uint64_t directly,
rather than passing it as a pointer.

Test: Built successfully with latest compilers.

Change-Id: Iecf67cb10a0f5e2b8ceb6363e0c4a9b927d1bfa9