seos: mask off discardable bit for all broadcast events
Split tid and evt into separate arguments to osTaskHandle (tid is
only passed to chre tasks)
Have the callers of osTaskHandle handle masking or not masking
the discardable bit (depending on if the event is private or broadcast)

Test: request and register for sensor event for gyro. send private event of
value 0x8206 (same value as what will be sent by gyro). Verify handler
receives the private event with event value 0x8206 and the sensor event with
value 0x0206.
Change-Id: I0c91edc105e8808d9455a84ba40c3c7b589cc3b3
Signed-off-by: Ben Fennema <fennema@google.com>