nanohubhal: Perform graceful shutdown in destructor
When destroying the NanoHub HAL singleton, perform graceful shutdown
including clean exit of the poll and app threads. This is the equivalent
cleanup done when a null callback is passed to subscribe_messages.

This fixes a crash seen during VTS testing where tests pass, but then
the default std::thread destructor is called for the NanoHub HAL
threads, which invokes std::terminate and causes the test to be marked
as failing.

Test: Context Hub VTS (currently under development)
Change-Id: Iaec2cad037ba81eaa19243d3019d81e433c9a2f8