bmi160: Set the min_step_buf to 2 for sensitive mode
Number of steps to suppress false positives at the begining of the walking
activity. After a walking activity has been detected, these steps will be
added to the total number of steps counted. As a consequence the step
counter output is suppressed during the first step of the walking activity
as defined in min_step_buf.

Bug: 32560999
Change-Id: I9c0ca981ab4c02dd34bf214549969db4d1126b1f