seos: fix external apps receiving unexpected evtData back
External apps calling eOsEnqueueEvt with a 3rd argument of 0 for the tid
don't expect to receive evtData back. (this usually occurs when
evtData is a value and not a pointer)
Change osEnqueueEvtAsApp's third argument to be a boolean indicating
whether the evtData value needs to be freed or not. If it doesn't need
to be freed, pass NULL for the pointer value to osEnqueueEvtCommon.
Change the checks for "don't need to be freed" to: isPtr && NULL
Fix spi/i2c currentTask in interrupt callback
Fix host interface currentTask

Test: confirm activity doesn't get its EVT_SENSOR_ACTIVITY and
SENSOR_DATA_EVENT_FLUSH values back in EVT_APP_FREE_EVT_DATA
Change-Id: I081c2064a5bf1c5ea7f4cc89ea0b8f3a377ea8f5
Signed-off-by: Ben Fennema <fennema@google.com>