Improved flush type-tracking for nanohub sensors.
In the case where two similar sensors are backed by the same physical
sensor (ie, gyro and uncalibrated gyro) nanohub is unaware of any
distinction between them.

As such, we previously had to guess at the true sensor type of each
flush event returned from nanohub. Instead, now we track the type of
each flush as it is requested so we can handle them correctly.

Change-Id: I7902198b1d0f16cedf08283e620a7bc00fd66bd8