firmware: stm32f4xx: Fix entry point in linker scripts
The entry point in the linker script is erroneously pointing to the
beginning of the vector table, while it should point to the reset
handler.
Since the entry point is discarded by objcopy during the following
conversion to raw binary, its wrong value has no impact to the
firmware written to and executed from flash.

The entry point is present in the elf file only.
If user loads the elf file in a debugger and restarts the execution,
debugger would jump to the elf entry point instead of using the value
in the vector table.

Fix the linker script using the reset handler as entry point.

Change-Id: I204451cbcc2c10395640f070e79794885ae2841b
Signed-off-by: Dmitry Shmidt <dimitrysh@google.com>