nanohub: set type NOLOAD to sections .stack and .codeplaceholder
The content of section .stack is unused, so no need to store it
in the ELF file nor to load it in memory when loading from ELF.
Currently the section's content appears in the ELF file because
it's marked with KEEP() in the linker script to avoid it being
removed by the garbage collector enabled by '--gc-sections'.

The section .codeplaceholder is just a trick to get in the BL
the starting address of the chre firmware. It's dummy content is
not used and can be removed from the ELF.

Add type 'NOLOAD' to these sections either to avoid including
their content in the ELF file and to specify that should be skip
when loading from ELF.

Change-Id: I05cea96d325aefc09a4478c69c8bbd2527188ae9
Signed-off-by: Antonio Borneo <antonio.borneo@st.com>