nanohub: fix BL's .data section
The initialization values for the BL's .data section should be in the
memory block 'bl' after BL code, not in the memory block 'code'.

Fix also objcopy command line to include the .data section, and remove
the reference to the non-existing .blrodata section.

Current BL code does not provide any content in the .data section that
is thus empty; this commit does not change the resulting binary output.

Change-Id: Ie03b5d6c39ad4937b8b198f0a6ee6b32189d1057
Signed-off-by: Antonio Borneo <antonio.borneo@st.com>