Make contexthub build with the VNDK.
When the vndk is enabled (BOARD_VNDK_VERSION=current),
vendor libraries must only link against vendor variants
in the build system. This also means that they receive
restricted sets of headers and that automatic global
headers are restricted.

A static library moved to being vendor is a build
time only change for libhubutilscommon.

Test: mma with BOARD_VNDK_VERSION=current
Bug: 69526027
Change-Id: I99573506afb3f91b846302be78a2bdcdb838d12c